統一數值型別為 Decimal，新增精度與小數位數屬性
移除 FieldDbType.Double，改為使用 FieldDbType.Decimal，統一處理十進位數值的邏輯，並新增 Precision 與 Scale 屬性以支援高精度需求。

更新 SQL 型別對應邏輯，動態生成 [decimal](precision, scale)，並調整資料庫型別解析邏輯以支援 Decimal 的精度與小數位數。

更新 DbType 和 Type 的對應邏輯，移除 Double 的處理，改為 Decimal 的對應。

新增 Decimal 型別的欄位比較與複製邏輯，並更新相關欄位描述與註解，明確說明屬性用途。

確保所有邏輯統一使用 Decimal，並移除與 Double 相關的處理。

+++ src/Bee.Base/Common/Common.cs
@@ -191,9 +191,9 @@ namespace Bee.Base
         /// </summary>
         Integer,
         /// <summary>
-        /// 浮點數。
+        /// 十進位數值（高精度）。
         /// </summary>
-        Double,
+        Decimal,
         /// <summary>
         /// 貨幣。
         /// </summary>
